:card_file_box: Reorganization in app fix #176
Add some folder and move files inside
