+++ src/App.tsx
@@ -10,22 +10,22 @@ import { ErrorContextProvider } from './context/error'
 import { ProfileContextProvider } from './context/profile'
 import Header from './components/Header'
 import Errors from './components/Errors'
-import PatientsPage from './pages/PatientsPage'
-import PrescriptionsPage from './pages/PrescriptionsPage'
-import PrescriptionEdit from './pages/PrescriptionEdit'
-import ProfilePage from './pages/ProfilePage'
-import PatientDetail from './pages/PatientDetail'
+import PatientsPage from './pages/patients/PatientsPage'
+import PrescriptionsPage from './pages/prescriptions/PrescriptionsPage'
+import PrescriptionEdit from './pages/prescriptions/PrescriptionEdit'
+import ProfilePage from './pages/accounts/ProfilePage'
+import PatientDetail from './pages/patients/PatientDetail'
 import PageNotFound from './pages/PageNotFound'
-import PatientEdit from './pages/PatientEdit'
-import PatientCreatePage from './pages/PatientCreatePage'
+import PatientEdit from './pages/patients/PatientEdit'
+import PatientCreatePage from './pages/patients/PatientCreatePage'
 import PrivateRoute from './components/PrivateRoute'
-import PrescriptionCreatePage from './pages/PrescriptionCreatePage'
-import LoginPage from './pages/LoginPage'
-import RegisterPage from './pages/RegisterPage'
-import AccountPage from './pages/AccountPage'
+import PrescriptionCreatePage from './pages/prescriptions/PrescriptionCreatePage'
+import LoginPage from './pages/authentifications/LoginPage'
+import RegisterPage from './pages/authentifications/RegisterPage'
+import AccountPage from './pages/accounts/AccountPage'
 import './App.css'
-import ResetPassword from './pages/ResetPassword'
-import NewResetPassword from './pages/NewResetPassword'
+import ResetPassword from './pages/authentifications/ResetPassword'
+import NewResetPassword from './pages/authentifications/NewResetPassword'
 
 library.add(fas, fab, far)
