+++ src/pages/patients/PatientsPage.tsx
@@ -1,13 +1,13 @@
-import PatientLine from '../components/PatientLine'
+import PatientLine from '../../components/PatientLine'
 import Container from 'react-bootstrap/Container'
 import Button from 'react-bootstrap/Button'
-import usePatients from '../hook/patient.hook'
-import useTranslationHook from '../hook/TranslationHook'
+import usePatients from '../../hook/patient.hook'
+import useTranslationHook from '../../hook/TranslationHook'
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
-import TitlePage from '../components/TitlePage'
+import TitlePage from '../../components/TitlePage'
 import { useCallback, useState } from 'react'
-import { Patient } from '../types'
-import SearchBar from '../components/SearchBar'
+import { Patient } from '../../types'
+import SearchBar from '../../components/SearchBar'
 
 const PatientsPage = (): JSX.Element => {
   const { patients, reloadPatients } = usePatients()
